refactor: split segment to additonal file
