+++ src/scenes/Analytics/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./most-played.component";
 export * from "./annual-sales.component";
 export * from "./popular-genres.component";
+export * from "./overall-stats.component";
